+++ addons/yat/src/YAT.cs
@@ -4,8 +4,10 @@ using YAT.Attributes;
 using YAT.Commands;
 using YAT.Helpers;
 using YAT.Interfaces;
+using YAT.Scenes.BaseTerminal;
 using YAT.Scenes.CommandManager;
-using YAT.Scenes.Terminal;
+using YAT.Scenes.EditorTerminal;
+using YAT.Scenes.GameTerminal;
 
 namespace YAT
 {
@@ -16,16 +18,9 @@ namespace YAT
 	public partial class YAT : Node
 	{
 		#region Signals
-		/// <summary>
-		/// Signal emitted when the YatOptions have been changed.
-		/// </summary>
-		/// <param name="options">The new YatOptions.</param>
 		[Signal]
 		public delegate void OptionsChangedEventHandler(YatOptions options);
 		[Signal]
-		/// <summary>
-		/// A signal that is emitted when the YAT addon is ready.
-		/// </summary>
 		public delegate void YatReadyEventHandler();
 		[Signal]
 		public delegate void TerminalOpenedEventHandler();
@@ -35,25 +30,24 @@ namespace YAT
 
 		[Export] public YatOptions Options { get; set; } = new();
 
-		public Terminal Terminal { get; private set; }
+		public BaseTerminal Terminal { get; private set; }
 		public Node Windows { get; private set; }
 		public Dictionary<string, ICommand> Commands { get; private set; } = new();
-
-		public readonly LinkedList<string> History = new();
-		public LinkedListNode<string> HistoryNode = null;
 
 		public OptionsManager OptionsManager { get; private set; }
 		public CommandManager CommandManager { get; private set; }
 
 		private bool _yatEnabled = true;
+		private GameTerminal _gameTerminal;
 
 		public override void _Ready()
 		{
 			CheckYatEnableSettings();
 
-			Terminal = GD.Load<PackedScene>("res://addons/yat/src/scenes/terminal/Terminal.tscn").Instantiate<Terminal>();
-			Terminal.Ready += () =>
+			_gameTerminal = GD.Load<PackedScene>("res://addons/yat/src/scenes/game_terminal/GameTerminal.tscn").Instantiate<GameTerminal>();
+			_gameTerminal.Ready += () =>
 			{
+				Terminal = _gameTerminal.BaseTerminal;
 				LogHelper.Terminal = Terminal;
 				OptionsManager.Load();
 
@@ -92,7 +86,8 @@ namespace YAT
 
 		public override void _UnhandledInput(InputEvent @event)
 		{
-			if (@event.IsActionPressed("yat_toggle")) ToggleTerminal();
+			if (!Engine.IsEditorHint() && @event.IsActionPressed("yat_toggle"))
+				ToggleTerminal();
 		}
 
 		/// <summary>
@@ -119,24 +114,24 @@ namespace YAT
 		{
 			if (!_yatEnabled) return;
 
-			if (Terminal.IsInsideTree()) CloseTerminal();
+			if (_gameTerminal.IsInsideTree()) CloseTerminal();
 			else OpenTerminal();
 		}
 
-		private void OpenTerminal()
+		public void OpenTerminal()
 		{
-			if (Terminal.IsInsideTree()) return;
+			if (_gameTerminal.IsInsideTree()) return;
 
-			Windows.AddChild(Terminal);
+			Windows.AddChild(_gameTerminal);
 
 			EmitSignal(SignalName.TerminalOpened);
 		}
 
-		private void CloseTerminal()
+		public void CloseTerminal()
 		{
-			if (!Terminal.IsInsideTree()) return;
+			if (!_gameTerminal.IsInsideTree()) return;
 
-			Windows.RemoveChild(Terminal);
+			Windows.RemoveChild(_gameTerminal);
 
 			EmitSignal(SignalName.TerminalClosed);
 		}