+++ addons/yat/src/commands/builtin/Options.cs
@@ -2,8 +2,8 @@ using Godot;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
-using YAT.Scenes.Terminal;
 using YAT.Scenes.YatWindow;
+using static YAT.Scenes.BaseTerminal.BaseTerminal;
 
 namespace YAT.Commands
 {
@@ -30,7 +30,7 @@ namespace YAT.Commands
 				return CommandResult.Success;
 			}
 
-			Yat.Terminal.Print("Options will be improved in the future.", Terminal.PrintType.Warning);
+			Yat.Terminal.Print("Options will be improved in the future.", PrintType.Warning);
 
 			_optionsWindowInstance = instanceValid ? _optionsWindowInstance : _optionsWindow.Instantiate<YatWindow>();
 			Yat.Windows.AddChild(_optionsWindowInstance);