+++ addons/yat/src/scenes/base_terminal/components/terminal_context/quic_commands_context/QuickCommandsContext.cs
@@ -2,7 +2,7 @@ using Godot;
 using YAT.Helpers;
 using YAT.Scenes.ContextMenu;
 
-namespace YAT.Scenes.Terminal
+namespace YAT.Scenes.BaseTerminal
 {
 	public partial class QuickCommandsContext : ContextSubmenu
 	{