+++ addons/yat/src/scenes/base_terminal/components/terminal_context/quick_commands/QuickCommands.cs
@@ -1,7 +1,7 @@
 using Godot;
 using Godot.Collections;
 
-namespace YAT.Scenes.Terminal
+namespace YAT.Scenes.BaseTerminal
 {
 	public partial class QuickCommands : Resource
 	{