+++ CHANGELOG.md
@@ -8,11 +8,15 @@ All notable changes to this project will be documented in this file.
 
 -   Scene helper class with PrintChildren & GetFromPathOrDefault methods.
 -   Print method overload to Terminal accepting StringBuilder.
+-   BaseTerminal scene.
+-   CloseRequested & TitleChanged signals to the BaseTerminal.
 
 ### Changed
 
 -   Simplified logic of ls command.
 -   Renamed TextHelper class to Text.
+-   Terminal renamed to GameTerminal.
+-   Terminal history is managed by BaseTerminal instead of YAT.
 
 ## [1.15.2-beta 2024-01-04]